requirements.txt
appdirs==1.4.3
Django==1.11
django-ckeditor==5.2.2
django-markdown-deux==1.0.5
django-pagedown==0.1.3
markdown2==2.3.4
packaging==16.8
Pillow==3.3.0
pkg-resources==0.0.0
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
